Add projects link to card title
