+++ src/components/ProjectsGrid.jsx
@@ -1,25 +1,5 @@
 import React from "react";
-import painting from "../assets/painting.jpg";
-
-const projects = [
-  {
-    title: "Proyecto 1",
-    description: "Descripción del proyecto",
-    image: painting,
-  },
-  {
-    title: "Proyecto 2",
-    description: "Descripción del proyecto",
-    image: painting,
-  },
-  {
-    title: "Proyecto 3",
-    description: "Descripción del proyecto",
-    image: painting,
-  },
-
-  // Aquí se van agregando más proyectos
-];
+import projects from "../data/projectsData";
 
 const ProjectsGrid = () => {
   return (
@@ -35,7 +15,11 @@ const ProjectsGrid = () => {
             className="w-full h-64 object-cover rounded-t-lg"
           />
           <div className="p-4 bg-black text-white font-sans rounded-b-lg">
-            <h2 className="text-xl font-bold">{project.title}</h2>
+            <a href={project.url} target="_blan" rel="noopener noreferrer">
+              <h2 className="text-xl font-bold hover:underline cursor-pointer">
+                {project.title}
+              </h2>
+            </a>
             <p className="mt-2">{project.description}</p>
           </div>
         </div>
