+++ src/data/projectsData.jsx
@@ -1,0 +1,25 @@
+import painting from "../assets/painting.jpg";
+
+const projects = [
+  {
+    title: "Proyecto 1",
+    description: "Descripción del Proyecto 1",
+    image: painting,
+    url: "https://link-al-proyecto-1.com",
+  },
+  {
+    title: "Proyecto 2",
+    description: "Descripción del Proyecto 2",
+    image: "/path/to/image2.jpg",
+    url: "https://link-al-proyecto-2.com",
+  },
+  {
+    title: "Proyecto 3",
+    description: "Descripción del Proyecto 3",
+    image: "/path/to/image3.jpg",
+    url: "https://link-al-proyecto-3.com",
+  },
+  // Aquí se van agregando más proyectos
+];
+
+export default projects;
